Updated animation example Fixed animation Added animation.gif to visually see the result in animation example's README.md Added custom FPS for each animation state (new variable for animation data) Changed Seconds to double from float

+++ examples/common/src/Systems/CollisionSystem.cpp
@@ -52,7 +52,7 @@ CollisionSystem::CollisionSystem()
 {
 }
 
-void CollisionSystem::update(float deltaTime)
+void CollisionSystem::update(double deltaTime)
 {
 	auto colliders = getEntities();
